feat(engine): implement monster behaviour and attack cooldown
This commit brings new functionality to the core.
It also fixes a few known issues, such as invalid attackSpeed
calculation. It also presents a new way to handle attack range
and speed, and adds mandatory attack cooldown for every entity.

All monsters are presented with the same behaviour. This may be changed
for some monsters in the feature, as the logic is defined in the base class.

Signed by Oles.