feat(gui): setup play/menu sessions and implement gui button
This commit adds another layer of abstraction for sessions.
The next thing to implement is SessionOrcehestrator,
which would manage the currently executing session
and switch to another screens when needed.

MainMenuSession (menu session) needs to be split up
in the future, as it contains generic implementation that
should be in a more general place, not in a session class.

Signed by Oles.
M--A *rainbow*